requirements.txt
alabaster==0.7.10
arrow==0.12.0
awscli==1.14.16
Babel==2.5.1
binaryornot==0.4.4
boto3==1.4.4
botocore==1.5.95
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
cookiecutter==1.6.0
coverage==4.4.2
docker==2.4.2
docker-pycreds==0.2.1
docutils==0.14
flake8==3.5.0
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.5
idna==2.6
imagesize==0.7.1
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
kenza-cli==0.0.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.13.3
poyo==0.4.1
pyasn1==0.4.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
PyYAML==3.12
requests==2.13.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.19.0
scipy==1.0.0
six==1.10.0
smmap2==2.0.3
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
spiono==0.0.1
urllib3==1.22
websocket-client==0.45.0
whichcraft==0.4.1
